issue 19: Refactored most of the work from RoboFile in to support classes.
